lab_8_SQL_FULL_ACCESS_PY/requirements.txt
# Lab 8 SQL FULL ACCESS (Python) - root / scripts
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
